edit task
edit task
	setup button/
	show add task view/
	retrieve task and show inside input boxes/
	update table/
	update view?/
		updated task

+++ app/src/main/java/edu/dlsu/mobapde/icandov3/DatabaseHelper.java
@@ -162,7 +162,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 
 
     //editTask
-    public boolean editTask(Task newtask, int currentId){
+    public boolean editTask(Task newtask, long currentId){
         SQLiteDatabase db = getWritableDatabase();
 
         ContentValues contentValues = new ContentValues();
